hafas-client@6, add vbb-stations@8 dev dependency, vbb-change-positions@2
hafas-client@6 doesn't have vbb-stations as a dependency anymore, which we had relied on implicitly, so we specify it explicitly now.

Because vbb-stations@8 uses IFOPT IDs (DHIDs), we update vbb-change-positions too in order to match them. Because we're waiting for https://github.com/juliuste/vbb-change-positions/pull/23 to be merged and published to npm, we temporarily resort to a Git-based dependency.